lib: Make React OnOff-Switch generally available
It's not just needed by one page anymore.
