+++ pkg/lib/cockpit-components-onoff.css
@@ -1,0 +1,6 @@
+/* on-off button styling depends on data-toggle in bootstrap */
+.btn-onoff-ct input {
+    position: absolute;
+    clip: rect(0, 0, 0, 0);
+    pointer-events: none;
+}
